Hyperlink styles and configuration.
Build number bump. Closes #25.

+++ ui-terminal/url_matcher.h
@@ -91,7 +91,7 @@ namespace ui {
         }; 
 
         bool isStateValid() const {
-            return static_cast<int>(state_) >= static_cast<int>(State::valid);
+            return state_ >= State::valid;
         }
 
         /** \name Character Groups
